Added Restaurant Functionality.
Added restaurant list, details, and edit functionality. The activites are a placeholder until the real activities are created. The app will list only the restaurants that the user created and with a long tap, the user can view the details of the restaurant and can edit. Delete functionality will be added next.